backend/requirements.txt
fastapi
uvicorn
pymongo
langchain
langchain-openai
langchain-mongodb
deepeval
openai
python-multipart
pydantic-settings
pypdf
python-docx
pandas
openpyxl
python-dotenv
aiofiles
pydub
